pyproject.toml
[tool.poetry]
name = "djate"
version = "0.1.0"
description = ""
authors = ["n1rjal <!-- <nirjalpaudel54312@gmail.com> -->"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
amqp = "5.2.0"
asgiref = "3.8.1"
attrs = "24.2.0"
billiard = "4.2.0"
celery = "5.4.0"
certifi = "2024.8.30"
click = "8.1.7"
click-didyoumean = "0.3.1"
click-plugins = "1.1.1"
click-repl = "0.3.0"
cron-descriptor = "1.4.5"
django = "5.1.1"
django-celery-beat = "2.7.0"
django-celery-results = "2.5.1"
django-cors-headers = "4.4.0"
django-stubs = "5.1.0"
django-stubs-ext = "5.1.0"
django-timezone-field = "7.0"
djangorestframework = "3.15.2"
drf-yasg = "1.21.7"
importlib-resources = "6.4.5"
inflection = "0.5.1"
jsonschema = "4.23.0"
jsonschema-specifications = "2023.12.1"
kombu = "5.4.0"
packaging = "24.1"
prompt-toolkit = "3.0.47"
psycopg2 = "2.9.9"
python-crontab = "3.2.0"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
pytz = "2024.1"
pyyaml = "6.0.2"
redis = "5.0.8"
referencing = "0.35.1"
rpds-py = "0.20.0"
sentry-sdk = "2.15.0"
setuptools = "74.1.2"
six = "1.16.0"
sqlparse = "0.5.1"
swagger-spec-validator = "3.0.4"
types-pyyaml = "6.0.12.20240917"
typing-extensions = "4.12.2"
tzdata = "2024.1"
uritemplate = "4.1.1"
urllib3 = "2.2.3"
vine = "5.1.0"
wcwidth = "0.2.13"

[tool.black]
line-length = 79
target-version = ['py311']
skip-string-normalization = false
extend-exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
  | migrations
)/
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3
lines_after_imports = 2
combine_as_imports = "true"
skip = [
    "__pycache__",
    ".venv",
    "migrations",
]

[tool.codespell]
skip = [
    "requirements.txt",
    "__pycache__",
    ".venv",
    "migrations",
    "staticfiles"
]